CPPCHECK fixes for printf issues
    Fix for following CPPCheck warnings
    - invalidPrintfArgType
    Rebased with latest develop branch

+++ openair1/PHY/LTE_TRANSPORT/dci_tools_common.c
@@ -996,7 +996,7 @@ uint8_t subframe2harq_pid(LTE_DL_FRAME_PARMS *frame_parms,uint32_t frame,uint8_t
     }
   }
 
-  AssertFatal(ret != 255, "invalid harq_pid(%d) at SFN/SF = %d/%d\n", (int8_t)ret, frame, subframe);
+  AssertFatal(ret != 255, "invalid harq_pid(%u) at SFN/SF = %u/%d\n", ret, frame, subframe);
   return ret;
 }
